(Re-upload) Added sql generator to pairing.php
I added the sql function to the pairing.php. I kept the parameters and added a new one, $pdo. It should also be sql injection protected and return a string.

Also I added player and winner to the construct function. It seem needed in the sql function but I don't know if it clashes with any of the requirements for the construct function. I'll review it in the get pairing pull request